# api-docs/api/swagger/swagger.yaml
openapi: 3.0.0
info:
    title: 'Guardian'
    description: 'The Guardian is a modular open-source solution that includes best-in-class identity management and decentralized ledger technology (DLT) libraries. At the heart of the Guardian solution is a sophisticated Policy Workflow Engine (PWE) that enables applications to offer a requirements-based tokenization implementation.'
    version: '1.2.1'
    contact:
        name: 'API developer'
        url: 'https://envisionblockchain.com'
        email: 'info@envisionblockchain.com'
    license:
        name: 'Apache 2.0'
        url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
    - url: /api/v1
      description: version 1.0
components:
    securitySchemes:
        bearerAuth:
            type: 'http'
            scheme: 'bearer'
            bearerFormat: 'JWT'
    schemas:
...
        User:
            type: object
            required:
                - username
                - role
            properties:
                confirmed:
                    type: string
                failed:
                    type: string
                username:
                    type: string
                role:
                    type: string
                hederaAccountId:
                    type: string
                hederaAccountKey:
                    type: string
                did:
                    type: string
                didDocument:
                    type: object
                vcDocument:
                    type: object
                addressBook:
                    type: object
                    properties:
                        appnetName:
                            type: string
                        addressBook:
                            type: string
                        didTopic:
                            type: string
                        vcTopic:
                            type: string
                        didServerUrl:
                            type: string
                        didTopicMemo:
                            type: string
                        vcTopicMemo:
                            type: string
...
        Error:
            type: object
            required:
                - code
                - message
            properties:
                code:
                    type: number
                message:
                    type: string
...
paths:
...
    /profiles/{username}/balance:
        get:
            tags:
                - profiles
            description: Requests Hedera account balance. Only users with the Installer role are allowed to make the request.
            parameters:
                - in: path
                  name: username
                  schema:
                      type: string
                  required: true
                  description: The name of the user for whom to fetch the balance.
            security:
                - bearerAuth: []
            summary: Returns user's Hedera account balance.
            responses:
                200:
                    description: Successful operation.
                    content:
                        application/json:
                            schema:
                                type: string
                401:
                    description: Unauthorized.
                403:
                    description: Forbidden.
                500:
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /profiles/{username}:
        get:
            tags:
                - profiles
            description: Returns user account information. For users with the Root Authority role it also returns address book and VC document information.
            parameters:
                - in: path
                  name: username
                  schema:
                      type: string
                  required: true
                  description: The name of the user for whom to fetch the information.
            security:
                - bearerAuth: []
            summary: Returns user account info.
            responses:
                200:
                    description: Successful operation.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
                401:
                    description: Unauthorized.
                403:
                    description: Forbidden.
                500:
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
        put:
            tags:
                - profiles
            description: Sets Hedera credentials for the user. For users with the Root Authority role it also creates an address book.
            parameters:
                - in: path
                  name: username
                  schema:
                      type: string
                  required: true
                  description: The name of the user for whom to update the information.
            security:
                - bearerAuth: []
            summary: Sets Hedera credentials for the user.
            requestBody:
                description: Object that contains the Hedera account data.
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/User'
            responses:
                201:
                    description: Successful operation.
                401:
                    description: Unauthorized.
                403:
                    description: Forbidden.
                500:
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
...